completed

+++ app/Http/Controllers/AdminController.php
@@ -9,6 +9,8 @@ use App\Models\Order;
 use App\Notifications\SendEmailNotification;
 use Barryvdh\DomPDF\Facade\PDF;
 use Illuminate\Support\Facades\Notification;
+use Illuminate\Support\Facades\Auth;
+
 
 
 
@@ -16,52 +18,82 @@ class AdminController extends Controller
 {
     public function view_catagory()
     {
-        $data = catagory::all();
-        return view('admin.catagory', compact('data'));
+        if (Auth::id()) {
+            $data = catagory::all();
+            return view('admin.catagory', compact('data'));
+        } else {
+            return redirect('login');
+        }
     }
 
     public function add_catagory(Request $request)
     {
-        $data = new catagory;
+        if (Auth::id()) {
 
-        $data->catagory_name = $request->catagory;
+            $data = new catagory;
 
-        $data->save();
+            $data->catagory_name = $request->catagory;
 
-        return redirect()->back()->with('message', 'Catagory added successfully');
+            $data->save();
+
+            return redirect()->back()->with('message', 'Catagory added successfully');
+        } else {
+            return redirect('login');
+        }
     }
 
     public function delete_catagory($id)
     {
-        $deldata = catagory::find($id);
-        $deldata->delete();
-        return redirect()->back()->with('message', 'Catagory Deleted ');
+        if (Auth::id()) {
+
+            $deldata = catagory::find($id);
+            $deldata->delete();
+            return redirect()->back()->with('message', 'Catagory Deleted ');
+        } else {
+            return redirect('login');
+        }
     }
+
 
     public function view_product()
     {
+        if (Auth::id()) {
 
-        $catagory = catagory::all();
-        return view('admin.product', compact('catagory'));
+            $catagory = catagory::all();
+            return view('admin.product', compact('catagory'));
+        }
+         else {
+            return redirect('login');
+        }
     }
 
     public function add_product(Request $request)
     {
-        $product = new products;
-        $product->title = $request->title;
-        $product->description = $request->description;
-        $product->price = $request->price;
-        $product->discount_price = $request->dis_price;
-        $product->quantity = $request->quantity;
+        if (Auth::id()) {
 
-        $product->catagory = $request->catagory;
+            $product = new products;
+            $product->title = $request->title;
+            $product->description = $request->description;
+            $product->price = $request->price;
+            $product->discount_price = $request->dis_price;
+            $product->quantity = $request->quantity;
 
-        $image = $request->image;
-        $imagename = time() . '.' . $image->getClientOriginalExtension();
-        $request->image->move('product_img', $imagename);
-        $product->image = $imagename;
-        $product->save();
-        return redirect()->back()->with('message', 'Product added successfully');
+            $product->catagory = $request->catagory;
+
+            $image = $request->image;
+            $imagename = time() . '.' . $image->getClientOriginalExtension();
+            $request->image->move('product_img', $imagename);
+            $product->image = $imagename;
+            $product->save();
+            return redirect()->back()->with('message', 'Product added successfully');
+        } 
+        else {
+            return redirect('login');
+        }
+
+        
+        
+      
     }
 
     public function show_product()
@@ -153,19 +185,17 @@ class AdminController extends Controller
     }
 
 
-    public function searchdata(Request $request){
+    public function searchdata(Request $request)
+    {
 
-        
-        $searchtxt= $request['search']?? "";
-        if($searchtxt!=""){
-            $order =order::where('name', 'LIKE', "%$searchtxt%")->orwhere('email', 'LIKE', "%$searchtxt%")->orwhere('phone', 'LIKE', "%$searchtxt%")->get();
-            
+
+        $searchtxt = $request['search'] ?? "";
+        if ($searchtxt != "") {
+            $order = order::where('name', 'LIKE', "%$searchtxt%")->orwhere('email', 'LIKE', "%$searchtxt%")->orwhere('phone', 'LIKE', "%$searchtxt%")->get();
+        } else {
+            $order = order::all();
         }
-        else{
-            $order=order::all();
-        }
-        $data=compact('order','searchtxt');
+        $data = compact('order', 'searchtxt');
         return view('admin.order')->with($data);
-        
     }
 }
